+++ inspector/src/components/actionTabs/lines/iconButtonLineComponent.tsx
@@ -4,6 +4,7 @@ export interface IIconButtonLineComponentProps {
     icon: string;
     onClick: () => void;
     tooltip: string;
+    active?: boolean;
 }
 
 export class IconButtonLineComponent extends React.Component<IIconButtonLineComponentProps> {
@@ -14,7 +15,7 @@ export class IconButtonLineComponent extends React.Component<IIconButtonLineComp
     render() {
 
         return (
-            <div title={this.props.tooltip} className={`icon ${this.props.icon}`} onClick={() => this.props.onClick()} />
+            <div style={{backgroundColor: this.props.active ? '#111111' : 'transparent'}} title={this.props.tooltip} className={`icon ${this.props.icon}`} onClick={() => this.props.onClick()} />
         );
     }
 }
